requirements.txt
# python==3.11.9

beartype==0.18.5
black==24.4.1
colorama==0.4.6
colorspacious==1.1.2
dacite==1.8.1
diffusers==0.27.2
dists-pytorch==0.1
e3nn==0.5.1
einops==0.7.0
hydra-core==1.3.2
imageio==2.34.1
imageio-ffmpeg==0.4.9
jaxtyping==0.2.28
lpips==0.1.4
matplotlib==3.8.4
moviepy==1.0.3
natsort==8.4.0
plyfile==1.0.3
pytorch-lightning==2.2.3
ruff==0.4.2
scikit-image==0.23.2
svg-py==1.4.3
tabulate==0.9.0
timm==0.9.16
tqdm==4.66.2
transformers==4.40.1
wandb==0.16.6
